requirements.txt
google-api-python-client
lxml
SQLAlchemy
youtube_dl
discord.py[voice]==1.3.3
pydub
aiohttp<3.7.0
ffmpeg-python